backend/requirements_fastapi.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.3.3
dnspython==2.3.0

# Authentication & Security
bcrypt==4.0.1
pyjwt==2.7.0

# HTTP & Utilities
requests==2.31.0
python-dotenv==1.0.0
